+++ src/Expectations/Time.php
@@ -106,12 +106,26 @@ expect()->extend(
 expect()->extend(
     'toBeLastYear',
     /**
-     * Assert the date is in the next year.
+     * Assert the date is in the last year.
      */
     function (): Expectation {
         $value = ValueProcessor::getCarbonDate($this->value);
 
         assertTrue($value->isLastYear(), sprintf('Failed to assert that [%s] is in the last year', $value));
+
+        return $this;
+    }
+);
+
+expect()->extend(
+    'toBeLastMonth',
+    /**
+     * Assert the date is in the last month.
+     */
+    function (): Expectation {
+        $value = ValueProcessor::getCarbonDate($this->value);
+
+        assertTrue($value->isLastMonth(), sprintf('Failed to assert that [%s] is in the last month', $value));
 
         return $this;
     }